Add new module and add checkOutTask function
