+++ src/modules/taskStatus.js
@@ -1,0 +1,28 @@
+import { getTasks, storeTask, render } from './functions.js';
+
+export const checkOutTask = (target, tasks) => {
+  if (target.classList.contains('list__task') || target.classList.contains('checkbox')) {
+    tasks = getTasks();
+
+    let targetInput;
+
+    if (target.classList.contains('list__task')) {
+      targetInput = target.firstChild.nextElementSibling.firstChild;
+    } else {
+      targetInput = target.nextElementSibling.firstChild;
+    }
+
+    const taskIndex = tasks.findIndex((task) => task.description === targetInput.value);
+
+    tasks[taskIndex].complete = !tasks[taskIndex].complete;
+
+    if (tasks[taskIndex].complete) {
+      tasks[taskIndex].checked = 'checked';
+    } else {
+      tasks[taskIndex].checked = '';
+    }
+
+    storeTask(tasks);
+    render(tasks);
+  }
+};
